Refactor schedule pages to bundled modules
