+++ racewayschedule.html
@@ -8,12 +8,7 @@
   <link rel="icon" href="icons/favicon.svg" type="image/svg+xml">
   <link rel="stylesheet" href="style.css">
   <script src="https://cdnjs.cloudflare.com/ajax/libs/xlsx/0.18.5/xlsx.full.min.js" defer></script>
-  <script src="dirtyTracker.js" defer></script>
-  <script type="module" src="site.js?v={{COMMIT_SHA}}"></script>
-  <script type="module" src="tableUtils.mjs?v={{COMMIT_SHA}}" defer></script>
-  <script type="module" src="e2e-helpers.js?v={{COMMIT_SHA}}"></script>
-  <script type="module" src="ductbankTable.js?v={{COMMIT_SHA}}"></script>
-  <script type="module" src="./racewayschedule.js?v={{COMMIT_SHA}}" defer></script>
+  <script type="module" src="dist/racewayschedule.js" defer></script>
 </head>
 <body>
   <a href="#main-content" class="skip-link">Skip to main content</a>
